Deleting only the current token that user is using to be authenticated

+++ app/Http/Controllers/AuthController.php
@@ -68,7 +68,7 @@ class AuthController extends Controller
 
     public function logout()
     {
-        Auth::user()->tokens()->delete();
+        Auth::user()->tokens()->where('name',config('auth.tokenName'))->delete();
 
         return response([
             'message' => 'Logged Out'
